bugfix: first approuch to the error:
when working with 3 axis simultaneously, after some minutes of working, moving to aleatory positions, the axis goes to Posicao_inicial instead of to Posicao_final.

In this case we are confirming at each moving, if the stop position is equivalent to the target position.

If not new attempts to reach the position is tried before raise an fatal error.